⚡️ Header에서 SearchBox 컴포넌트 분리
